+++ .gitmodules
@@ -1,3 +1,4 @@
 [submodule "deps/finch-core"]
 	path = deps/finch-core
-	url = https://github.com/runfinch/finch-core.git
+	url = https://github.com/ginglis13/finch-core.git
+	branch = windev
